put the comments function in a component
